ir/program: splits program into its own module
